requirements.txt
Django>=4.2,<5.2
whitenoise==6.6.0
gunicorn==21.2.0
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2
python-decouple==3.8